+++ packages/contracts/src/index.ts
@@ -1,0 +1,14 @@
+export {
+  HudEventSchema,
+  HudEventTypes,
+  type HudEvent,
+  type HudEventType,
+  type SessionStartEvent,
+  type SessionEndEvent,
+  type PromptSubmitEvent,
+  type ToolUseEvent,
+  type TurnStopEvent,
+  type CompactStartEvent,
+  type CompactEndEvent,
+  type HudErrorEvent,
+} from './event';